requirements.txt
streamlit
streamlit-lottie
streamlit-option-menu
firebase-admin
lottie
Pillow
numpy
tensorflow
requests
plotly
matplotlib
seaborn
langchain
torch
accelerate
transformers
sentence_transformers
streamlit_chat
faiss-cpu
altair
tiktoken
huggingface-hub
pypdf
ctransformers
pymupdf
llama-cpp-python

